Refactoring-code
