+++ src/main/java/com/networknt/schema/EnumValidator.java
@@ -32,7 +32,6 @@ public class EnumValidator extends BaseJsonValidator implements JsonValidator {
 
     private final Set<JsonNode> nodes;
     private final String error;
-    private final ValidationContext validationContext;
 
     public EnumValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext) {
         super(schemaPath, schemaNode, parentSchema, ValidatorTypeCode.ENUM, validationContext);
