+++ src/main/java/com/networknt/schema/FormatValidator.java
@@ -29,7 +29,6 @@ public class FormatValidator extends BaseJsonValidator implements JsonValidator 
     private static final Logger logger = LoggerFactory.getLogger(FormatValidator.class);
 
     private final Format format;
-    private final ValidationContext validationContext;
 
     public FormatValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext, Format format) {
         super(schemaPath, schemaNode, parentSchema, ValidatorTypeCode.FORMAT, validationContext);
