+++ src/main/java/com/networknt/schema/JsonMetaSchema.java
@@ -93,138 +93,6 @@ public class JsonMetaSchema {
         COMMON_BUILTIN_FORMATS.add(pattern("style", "\\s*(.+?):\\s*([^;]+);?"));
     }
 
-    private static class V4 {
-        private static String URI = "https://json-schema.org/draft-04/schema";
-        private static final String ID = "id";
-
-        public static final List<Format> BUILTIN_FORMATS = new ArrayList<Format>(JsonMetaSchema.COMMON_BUILTIN_FORMATS);
-
-        static {
-            // add version specific formats here.
-            //BUILTIN_FORMATS.add(pattern("phone", "^\\+(?:[0-9] ?){6,14}[0-9]$"));
-        }
-
-        public static JsonMetaSchema getInstance() {
-            return new Builder(URI)
-                    .idKeyword(ID)
-                    .addFormats(BUILTIN_FORMATS)
-                    .addKeywords(ValidatorTypeCode.getNonFormatKeywords(SpecVersion.VersionFlag.V4))
-                    // keywords that may validly exist, but have no validation aspect to them
-                    .addKeywords(Arrays.asList(
-                            new NonValidationKeyword("$schema"),
-                            new NonValidationKeyword("id"),
-                            new NonValidationKeyword("title"),
-                            new NonValidationKeyword("description"),
-                            new NonValidationKeyword("default"),
-                            new NonValidationKeyword("definitions"),
-                            new NonValidationKeyword("exampleSetFlag")
-                    ))
-                    .build();
-        }
-    }
-
-    private static class V6 {
-        private static String URI = "https://json-schema.org/draft-06/schema";
-        // Draft 6 uses "$id"
-        private static final String ID = "$id";
-
-        public static final List<Format> BUILTIN_FORMATS = new ArrayList<Format>(JsonMetaSchema.COMMON_BUILTIN_FORMATS);
-
-        static {
-            // add version specific formats here.
-            //BUILTIN_FORMATS.add(pattern("phone", "^\\+(?:[0-9] ?){6,14}[0-9]$"));
-        }
-
-        public static JsonMetaSchema getInstance() {
-            return new Builder(URI)
-                    .idKeyword(ID)
-                    .addFormats(BUILTIN_FORMATS)
-                    .addKeywords(ValidatorTypeCode.getNonFormatKeywords(SpecVersion.VersionFlag.V6))
-                    // keywords that may validly exist, but have no validation aspect to them
-                    .addKeywords(Arrays.asList(
-                            new NonValidationKeyword("$schema"),
-                            new NonValidationKeyword("$id"),
-                            new NonValidationKeyword("title"),
-                            new NonValidationKeyword("description"),
-                            new NonValidationKeyword("default"),
-                            new NonValidationKeyword("definitions")
-                    ))
-                    .build();
-        }
-    }
-
-    private static class V7 {
-        private static String URI = "https://json-schema.org/draft-07/schema";
-        private static final String ID = "$id";
-
-        public static final List<Format> BUILTIN_FORMATS = new ArrayList<Format>(JsonMetaSchema.COMMON_BUILTIN_FORMATS);
-
-        static {
-            // add version specific formats here.
-            //BUILTIN_FORMATS.add(pattern("phone", "^\\+(?:[0-9] ?){6,14}[0-9]$"));
-        }
-
-        public static JsonMetaSchema getInstance() {
-            return new Builder(URI)
-                    .idKeyword(ID)
-                    .addFormats(BUILTIN_FORMATS)
-                    .addKeywords(ValidatorTypeCode.getNonFormatKeywords(SpecVersion.VersionFlag.V7))
-                    // keywords that may validly exist, but have no validation aspect to them
-                    .addKeywords(Arrays.asList(
-                            new NonValidationKeyword("$schema"),
-                            new NonValidationKeyword("$id"),
-                            new NonValidationKeyword("title"),
-                            new NonValidationKeyword("description"),
-                            new NonValidationKeyword("default"),
-                            new NonValidationKeyword("definitions"),
-                            new NonValidationKeyword("$comment"),
-                            new NonValidationKeyword("contentMediaType"),
-                            new NonValidationKeyword("contentEncoding"),
-                            new NonValidationKeyword("examples"),
-                            new NonValidationKeyword("message")
-                    ))
-                    .build();
-        }
-    }
-
-    private static class V201909 {
-        private static String URI = "https://json-schema.org/draft/2019-09/schema";
-        private static final String ID = "$id";
-
-        public static final List<Format> BUILTIN_FORMATS = new ArrayList<Format>(JsonMetaSchema.COMMON_BUILTIN_FORMATS);
-
-        static {
-            // add version specific formats here.
-            //BUILTIN_FORMATS.add(pattern("phone", "^\\+(?:[0-9] ?){6,14}[0-9]$"));
-        }
-
-        public static JsonMetaSchema getInstance() {
-            return new Builder(URI)
-                    .idKeyword(ID)
-                    .addFormats(BUILTIN_FORMATS)
-                    .addKeywords(ValidatorTypeCode.getNonFormatKeywords(SpecVersion.VersionFlag.V201909))
-                    // keywords that may validly exist, but have no validation aspect to them
-                    .addKeywords(Arrays.asList(
-                            new NonValidationKeyword("$schema"),
-                            new NonValidationKeyword("$id"),
-                            new NonValidationKeyword("title"),
-                            new NonValidationKeyword("description"),
-                            new NonValidationKeyword("default"),
-                            new NonValidationKeyword("definitions"),
-                            new NonValidationKeyword("$comment"),
-                            new NonValidationKeyword("$defs"),  // newly added in 2019-09 release.
-                            new NonValidationKeyword("$anchor"),
-                            new NonValidationKeyword("additionalItems"),
-                            new NonValidationKeyword("deprecated"),
-                            new NonValidationKeyword("contentMediaType"),
-                            new NonValidationKeyword("contentEncoding"),
-                            new NonValidationKeyword("examples"),
-                            new NonValidationKeyword("then")
-                    ))
-                    .build();
-        }
-    }
-
     public static class Builder {
         private Map<String, Keyword> keywords = new HashMap<String, Keyword>();
         private Map<String, Format> formats = new HashMap<String, Format>();
@@ -312,19 +180,19 @@ public class JsonMetaSchema {
     }
 
     public static JsonMetaSchema getV4() {
-        return V4.getInstance();
+        return new Version4().getInstance();
     }
 
     public static JsonMetaSchema getV6() {
-        return V6.getInstance();
+        return new Version6().getInstance();
     }
 
     public static JsonMetaSchema getV7() {
-        return V7.getInstance();
+        return new Version7().getInstance();
     }
 
     public static JsonMetaSchema getV201909() {
-        return V201909.getInstance();
+        return new Version201909().getInstance();
     }
 
     /**
