+++ src/main/java/com/networknt/schema/MinimumValidator.java
@@ -37,8 +37,6 @@ public class MinimumValidator extends BaseJsonValidator implements JsonValidator
      */
     private final ThresholdMixin typedMinimum;
 
-    private final ValidationContext validationContext;
-
     public MinimumValidator(String schemaPath, final JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext) {
         super(schemaPath, schemaNode, parentSchema, ValidatorTypeCode.MINIMUM, validationContext);
