+++ src/main/java/com/networknt/schema/PropertiesValidator.java
@@ -29,7 +29,6 @@ public class PropertiesValidator extends BaseJsonValidator implements JsonValida
     public static final String PROPERTY = "properties";
     private static final Logger logger = LoggerFactory.getLogger(PropertiesValidator.class);
     private final Map<String, JsonSchema> schemas = new HashMap<String, JsonSchema>();
-    private final ValidationContext validationContext;
 
     public PropertiesValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext) {
         super(schemaPath, schemaNode, parentSchema, ValidatorTypeCode.PROPERTIES, validationContext);
