+++ src/main/java/com/networknt/schema/UUIDValidator.java
@@ -31,7 +31,6 @@ public class UUIDValidator extends BaseJsonValidator implements JsonValidator {
 
 
     private static final String regex = "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$";
-    private final ValidationContext validationContext;
 
     public UUIDValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext, String formatName) {
         super(schemaPath, schemaNode, parentSchema, ValidatorTypeCode.UUID, validationContext);
