+++ src/main/java/com/networknt/schema/Version6.java
@@ -1,0 +1,31 @@
+package com.networknt.schema;
+
+import java.util.Arrays;
+
+public class Version6 extends JsonSchemaVersion{
+    private static String URI = "https://json-schema.org/draft-06/schema";
+    // Draft 6 uses "$id"
+    private static final String ID = "$id";
+
+    static {
+        // add version specific formats here.
+        //BUILTIN_FORMATS.add(pattern("phone", "^\\+(?:[0-9] ?){6,14}[0-9]$"));
+    }
+
+    JsonMetaSchema getInstance() {
+        return new JsonMetaSchema.Builder(URI)
+                .idKeyword(ID)
+                .addFormats(BUILTIN_FORMATS)
+                .addKeywords(ValidatorTypeCode.getNonFormatKeywords(SpecVersion.VersionFlag.V6))
+                // keywords that may validly exist, but have no validation aspect to them
+                .addKeywords(Arrays.asList(
+                        new NonValidationKeyword("$schema"),
+                        new NonValidationKeyword("$id"),
+                        new NonValidationKeyword("title"),
+                        new NonValidationKeyword("description"),
+                        new NonValidationKeyword("default"),
+                        new NonValidationKeyword("definitions")
+                ))
+                .build();
+    }
+}
