+++ src/main/java/com/networknt/schema/Version7.java
@@ -1,0 +1,35 @@
+package com.networknt.schema;
+
+import java.util.Arrays;
+
+public class Version7 extends JsonSchemaVersion{
+    private static String URI = "https://json-schema.org/draft-07/schema";
+    private static final String ID = "$id";
+
+    static {
+        // add version specific formats here.
+        //BUILTIN_FORMATS.add(pattern("phone", "^\\+(?:[0-9] ?){6,14}[0-9]$"));
+    }
+    @Override
+    JsonMetaSchema getInstance() {
+        return new JsonMetaSchema.Builder(URI)
+                .idKeyword(ID)
+                .addFormats(BUILTIN_FORMATS)
+                .addKeywords(ValidatorTypeCode.getNonFormatKeywords(SpecVersion.VersionFlag.V7))
+                // keywords that may validly exist, but have no validation aspect to them
+                .addKeywords(Arrays.asList(
+                        new NonValidationKeyword("$schema"),
+                        new NonValidationKeyword("$id"),
+                        new NonValidationKeyword("title"),
+                        new NonValidationKeyword("description"),
+                        new NonValidationKeyword("default"),
+                        new NonValidationKeyword("definitions"),
+                        new NonValidationKeyword("$comment"),
+                        new NonValidationKeyword("contentMediaType"),
+                        new NonValidationKeyword("contentEncoding"),
+                        new NonValidationKeyword("examples"),
+                        new NonValidationKeyword("message")
+                ))
+                .build();
+    }
+}
